pyproject.toml
[tool.poetry]
name = "serial-pc-control"
version = "0.1.0"
description = "Control your PC keyboard and mouse via serial messages."
authors = ["carlosperate <carlosperate@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.6"
pyserial = "^3.4"
pyautogui = "^0.9.48"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
